Move merge into utils
